requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
awsebcli==3.21.0
billiard==4.2.1
blessed==1.20.0
boto3==1.26.137
botocore==1.35.99
celery==5.4.0
cement==2.10.14
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.1.3
django-autoslug==1.9.9
django-axes==7.0.0
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-crontab==0.7.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-recaptcha==4.0.0
django-ses==4.2.0
django-sslserver==0.22
django-storages==1.14.4
django-taggit==6.1.0
django-templated-mail==1.1.1
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
fire==0.7.0
frozenlist==1.5.0
geoip2==4.8.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==2.6.2
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
pathspec==0.10.1
pillow==11.0.0
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.21
Pygments==2.18.0
PyJWT==2.7.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.3
s3transfer==0.6.1
semantic-version==2.10.0
setuptools==75.4.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
termcolor==2.5.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.1
yarl==1.17.1
